pyproject.toml
[tool.poetry]
name = "Axiom"
version = "0.1.0"
description = "A utility library for internal use"
readme = "README.md"
authors = ["Thomas Genet <thomas.genet@mosaic.com>"]
license = "MIT"
repository = "https://github.com/mosaicoperations/axiom-ops"
documentation = "https://github.com/CarrierOps/1P-Wiki/tree/main/ByteSizedLearning/Description%20Of%20Services"
packages = [
    { include = "axiom", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10"
google-cloud-logging = "^3.11.0"
google-cloud-storage = "^2.18.0"
google-cloud-bigquery = "^3.25.0"
requests = "^2.32.3"
pyyaml = "^6.0.1"
office365-rest-python-client = "^2.5.11"
pandas = "^2.2.2"
polars = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = ">=6.0"
black = ">=20.8b1"
isort = ">=5.7.0"
mypy = ">=0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py37"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.7"
strict = true
